add entity
