+++ backend/docker-compose.local.yml
@@ -6,7 +6,7 @@ services:
     build:
       context: .
       dockerfile: docker/app/Dockerfile
-    command: uvicorn main:app --reload --host 0.0.0.0 --port 8080
+    command: uvicorn app.main:app --reload --host 0.0.0.0 --port 8080
     ports:
     - 8080:8080
     volumes:
