+++ backend/docker/app/Dockerfile
@@ -19,4 +19,4 @@ RUN pip install --no-cache-dir -r /requirements.txt
 
 COPY . /app
 
-CMD uvicorn main:app --host 0.0.0.0 --port $PORT
+CMD uvicorn app.main:app --host 0.0.0.0 --port $PORT
