--- backend/heroku.yml
@@ -1,6 +1,0 @@
-build:
-  docker:
-    web: docker/app/Dockerfile
-  run:
-    web: uvicorn main:app --host 0.0.0.0 --port 80
-
