+++ backend/requirements.txt
@@ -3,3 +3,4 @@ uvicorn==0.12.2
 # pandas
 psycopg2
 SQLAlchemy==1.3.20
+alembic
